Move search apis into separate files.
