+++ bbs/src/apis/common.ts
@@ -3,12 +3,8 @@ import { AxiosProgressEvent } from 'axios'
 import { UploadResponse } from '@/common/interfaces/base'
 import { Forum, ForumDetails } from '@/common/interfaces/forum'
 import {
-  GenericList,
   IndexData,
-  SearchSummaryResponse,
-  SearchSummaryUser,
   Thread,
-  ThreadInList,
   ThreadList,
   TopList,
 } from '@/common/interfaces/response'
@@ -101,69 +97,6 @@ export const getIndexData = async ({
   })
   result.top_list && transformTopList(result.top_list)
   return result
-}
-
-export const searchSummary = async (query: string) => {
-  const result = await request.get<SearchSummaryResponse>(
-    `${commonUrl}/search/summary`,
-    {
-      params: { q: query },
-    }
-  )
-  result.threads?.forEach(
-    (item) =>
-      (item.subject = unescapeSubject(item.subject, item.dateline, true))
-  )
-  if (result.tid_match) {
-    result.tid_match.subject = unescapeSubject(
-      result.tid_match.subject,
-      result.tid_match.dateline,
-      true
-    )
-  }
-  return result
-}
-
-export const searchThreads = ({
-  keyword,
-  author,
-  digest,
-  page,
-}: {
-  keyword?: string | null
-  author?: string | null
-  digest?: boolean
-  page?: number
-}) => {
-  return request.get<GenericList<ThreadInList>>(`${commonUrl}/search/threads`, {
-    params: {
-      ...(keyword && { q: keyword }),
-      ...(author && { author }),
-      ...(digest && { digest: 1 }),
-      ...(page && { page }),
-    },
-  })
-}
-
-export const searchUsers = ({
-  query,
-  withFriends,
-  page,
-}: {
-  query?: string | null
-  withFriends?: boolean
-  page?: number
-}) => {
-  return request.get<GenericList<SearchSummaryUser>>(
-    `${commonUrl}/search/users`,
-    {
-      params: {
-        ...(query && { q: query }),
-        ...(withFriends && { with_friends: 1 }),
-        ...(page && { page }),
-      },
-    }
-  )
 }
 
 export const getThreadList = async (params: {
